230303 도중

+++ DirectX3D/Framework/Math/Vector2.h
@@ -7,7 +7,7 @@ struct Vector2
 
     Vector2() = default;
     Vector2(float x, float y) : x(x), y(y) {};
-    Vector2(POINT point) : x(point.x), y(point.y) {};
+    Vector2(POINT point) : x((float)point.x), y((float)point.y) {};
     Vector2(Float2 point) : x(point.x), y(point.y) {};
 
     Vector2 operator+(const Vector2& value) const
